+++ lib/parser.dart
@@ -49,3 +49,4 @@ part 'src/parser/tok_str_parser.dart';
 part 'src/parser/tuple_parser.dart';
 part 'src/parser/val_parser.dart';
 part 'src/parser/white_parser.dart';
+part 'src/parser/expected_parser.dart';