+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.9
 
 - Minor cosmetic improvements in the `json_parser.dart` example file.
+- Added parser `ExpectedParser`
 - Added parser `WhiteParser`
 
 ## 0.1.8